% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{adjust.snet}
\alias{adjust.snet}
\title{Geodetic control network adjustment and design}
\usage{
adjust.snet(
  adjust = TRUE,
  survey.net,
  dim_type = list("1D", "2D"),
  sd.apriori = 1,
  wdh_model = list("n_dh", "sd_dh", "d_dh", "E"),
  n0 = 1,
  maxiter = 50,
  prob = 0.95,
  output = list("spatial", "report"),
  coord_tolerance = 0.001,
  result.units = list("mm", "cm", "m"),
  ellipse.scale = 1,
  teta.unit = list("deg", "rad"),
  units.dir = "sec",
  use.sd.estimated = TRUE,
  all = TRUE
)
}
\arguments{
\item{adjust}{logical. If TRUE adjustment will be performed. If FALSE the computation of the control network design quality will be conducted, Default: TRUE}

\item{survey.net}{surveynet object. Output from \code{read_surveynet} function}

\item{dim_type}{Type of the geodetic control network. Can be either 1D or 2D, Default: list("1D", "2D")}

\item{sd.apriori}{Apriori dispersion factor, Default: 1}

\item{wdh_model}{Weighted model for leveling measurements, Default: list("n_dh", "sd_dh", "d_dh", "E")}

\item{n0}{Number of station in the reference measurements (only for 1D network), Default: 1}

\item{maxiter}{Maximum number of iterations in adjustment computation, Default: 50}

\item{prob}{Probability in statistical testing, Default: 0.95}

\item{output}{Type of the output. It can be either, \code{spatial} (\code{sf} classes) or \code{report} (data frame), Default: list("spatial", "report")}

\item{coord_tolerance}{Tolerance in coordinate difference in two iteration, Default: 0.001}

\item{result.units}{Units of the results, Default: list("mm", "cm", "m")}

\item{ellipse.scale}{Scale parameter for displaying absolute error ellipses, Default: 1}

\item{teta.unit}{Units for orientation angle of the error ellipses, Default: list("deg", "rad")}

\item{units.dir}{Units for the residuals of the angular measurements, Default: 'sec'}

\item{use.sd.estimated}{logical, if estimated reference standard deviation factor should be used, Default: TRUE}

\item{all}{logical, if specific computation matrices should be exported, Default: FALSE}
}
\value{
if \code{outout = spatial} a list of three sf classes is exported. Otherwise \code{dataframe} object are exported.
}
\description{
Adjustment and design computation of the 1D and 2D geodetic control network.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[sf]{st_crs}},\code{\link[sf]{st_as_sf}},\code{\link[sf]{sfc}},\code{\link[sf]{sf}},\code{\link[sf]{st_geometry}}
\code{\link[tidyr]{pivot_longer}},\code{\link[tidyr]{separate}},\code{\link[tidyr]{pivot_wider}}
\code{\link[purrr]{when}},\code{\link[purrr]{keep}}
\code{\link[dplyr]{select}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{arrange}},\code{\link[dplyr]{filter}},\code{\link[dplyr]{mutate-joins}},\code{\link[dplyr]{mutate_all}},\code{\link[dplyr]{rename}}
\code{\link[stringr]{str_c}}
\code{\link[MASS]{ginv}}
}
