\name{allNA}
\alias{allNA}
\title{Tests for all NA values.}
\usage{
  allNA(x)
}
\arguments{
  \item{x}{A list, data frame or vector}
}
\value{
  TRUE if all values are NA, FALSE otherwise
}
\description{
  Tests for all NA values.
}
\seealso{
  \code{\link{as.surveyorStats}}

  Other stats helper functions: \code{\link{allNull}},
  \code{\link{is.yesno}}, \code{\link{qType}},
  \code{\link{reorderQuestion}},
  \code{\link{reorderResponse}},
  \code{\link{splitBinCombine}},
  \code{\link{splitMeanCombine}},
  \code{\link{splitPercentCombine}},
  \code{\link{weightedCount}}, \code{\link{weightedMean}},
  \code{\link{weightedMedian}}, \code{\link{weightedSum}}
}
\keyword{internal}

