\name{allNull}
\alias{allNull}
\title{Tests for all NULL values}
\usage{
  allNull(x)
}
\arguments{
  \item{x}{A list, data frame or vector}
}
\value{
  TRUE if all values are NULL, FALSE otherwise
}
\description{
  Tests for all NULL values
}
\seealso{
  \code{\link{as.surveyorStats}}

  Other stats helper functions: \code{\link{allNA}},
  \code{\link{is.yesno}}, \code{\link{qType}},
  \code{\link{reorderQuestion}},
  \code{\link{reorderResponse}},
  \code{\link{splitBinCombine}},
  \code{\link{splitMeanCombine}},
  \code{\link{splitPercentCombine}},
  \code{\link{weightedCount}}, \code{\link{weightedMean}},
  \code{\link{weightedMedian}}, \code{\link{weightedSum}}
}
\keyword{internal}

