% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{Amat}
\alias{Amat}
\title{FUNCTION_TITLE}
\usage{
Amat(survey.net, units)
}
\arguments{
\item{survey.net}{PARAM_DESCRIPTION}

\item{units}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}},\code{\link[dplyr]{select}}
\code{\link[tidyr]{spread}}
}
