% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{Amat1D}
\alias{Amat1D}
\title{FUNCTION_TITLE}
\usage{
Amat1D(survey.net)
}
\arguments{
\item{survey.net}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
