\name{as.surveyor_code}
\alias{as.surveyor_code}
\title{Creates object of class "surveyor_code".}
\usage{
  as.surveyor_code(x, surveyor, qid)
}
\arguments{
  \item{x}{Data created by surveyor code function}

  \item{surveyor}{A surveyor object}

  \item{qid}{Question identifier, e.g. "Q4"}
}
\value{
  An object of class "surveyor_code"
}
\description{
  Creates object of class "surveyor_code".
}

