\name{as.surveyorCode}
\alias{as.surveyorCode}
\title{Creates object of class "surveyorCode".}
\usage{
  as.surveyorCode(x, surveyor, qid, ...)
}
\arguments{
  \item{x}{Data created by surveyor code function}

  \item{surveyor}{A surveyor object}

  \item{qid}{Question identifier, e.g. "Q4"}

  \item{...}{Passed to surveyorStats}
}
\value{
  An object of class "surveyorCode"
}
\description{
  Creates object of class "surveyorCode".
}

