\name{as_surveyor_plot}
\alias{as_surveyor_plot}
\title{Creates surveyor_plot object.}
\usage{as_surveyor_plot(plot, expansion=1, plot_function="")}
\description{Creates surveyor_plot object.}
\details{Creates surveyor_plot object, a container for either ggplot or lattice graphic.}
\value{A surveyor_plot object}
\keyword{internal}
\arguments{\item{plot}{A ggplot or lattice object}
\item{expansion}{Multiplier for plot vertical dimension}
\item{plot_function}{The plot function that was used to create the plot}}
