\name{code_array_fast}
\alias{code_array_fast}
\title{Code survey data in multiple question form...}
\usage{code_array_fast(surveyor, q_id, remove_other=FALSE)}
\description{Code survey data in multiple question form}
\details{Code survey data in multiple question form (i.e. with subquestions)}
\value{data frame}
\seealso{Coding functions:
\code{\link{code_single}}, 
\code{\link{code_array}},
\code{\link{code_text}}
Summarising functions:
\code{\link{stats_bin}}, 
\code{\link{stats_net_score}}
Plot functions: 
\code{\link{plot_bar}}, 
\code{\link{plot_point}}}
\seealso{Other coding functions: 
\code{\link{code_single}},  
\code{\link{code_array}}, 
\code{\link{code_text}}, 
\code{\link{code_guess}}

For an overview of the surveyor package \code{\link{surveyor}}}
\value{data frame}
\keyword{code}
\arguments{\item{surveyor}{Surveyor object}
\item{q_id}{Question id}
\item{remove_other}{If true, will remove last column}
\item{surveyor}{Surveyor object}
\item{q_id}{Question id}
\item{remove_other}{If true, will remove last column}}
