\name{code_guess}
\alias{code_guess}
\title{Code survey data in single or array question form...}
\usage{code_guess(surveyor, q_id, ...)}
\description{Code survey data in single or array question form}
\details{Code survey data in array question form (i.e. with subquestions)}
\value{data frame}
\seealso{Coding functions:
\code{\link{code_single}}, 
\code{\link{code_array}},
\code{\link{code_text}}
Summarising functions:
\code{\link{stats_bin}}, 
\code{\link{stats_net_score}}
Plot functions: 
\code{\link{plot_bar}}, 
\code{\link{plot_point}}}
\arguments{\item{surveyor}{Surveyor object}
\item{q_id}{Question id}
\item{...}{Other parameters passed on to downstream code_* functions}}
