\name{code_rank}
\alias{code_rank}
\title{Code survey data in rank question form (i...}
\usage{code_rank(surveyor, Qid, remove.other=FALSE)}
\description{Code survey data in rank question form (i.e. with subquestions)}
\seealso{code_single, code_array}
\arguments{\item{surveyor}{Surveyor object}
\item{Qid}{Question id}
\item{remove.other}{Boolean}}
