\name{code_single}
\alias{code_single}
\title{Code survey data in array question form...}
\usage{code_single(surveyor, q_id)}
\description{Code survey data in array question form}
\details{Code survey data in array question form (i.e. with subquestions)}
\seealso{Coding functions:
\code{\link{code_single}}, 
\code{\link{code_array}},
\code{\link{code_text}}
Summarising functions:
\code{\link{stats_bin}}, 
\code{\link{stats_net_score}}
Plot functions: 
\code{\link{plot_bar}}, 
\code{\link{plot_point}}}
\value{data frame}
\arguments{\item{surveyor}{Surveyor object}
\item{q_id}{Question id}}
