\name{code_text}
\alias{code_text}
\title{Code survey data in text question form (i...}
\usage{code_text(surveyor, Qid, inn, out)}
\description{Code survey data in text question form (i.e. open-ended text}
\seealso{code_single, code_rank}
\arguments{\item{surveyor}{Surveyor object}
\item{Qid}{Question id}
\item{inn}{string passed to latexTranslate()}
\item{out}{string passed to latexTranslate()}}
