% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{coef_d}
\alias{coef_d}
\title{FUNCTION_TITLE}
\usage{
coef_d(pt1, pt2, pts, units)
}
\arguments{
\item{pt1}{PARAM_DESCRIPTION}

\item{pt2}{PARAM_DESCRIPTION}

\item{pts}{PARAM_DESCRIPTION}

\item{units}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
