% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{dec2dms}
\alias{dec2dms}
\title{dec2dms}
\usage{
dec2dms(ang)
}
\arguments{
\item{ang}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
