% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{dist}
\alias{dist}
\title{dist}
\usage{
dist(pt1_coords, pt2_coords)
}
\arguments{
\item{pt1_coords}{PARAM_DESCRIPTION}

\item{pt2_coords}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Calculate distance
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
