% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{error.ellipse}
\alias{error.ellipse}
\title{FUNCTION_TITLE}
\usage{
error.ellipse(Qxy, prob = NA, sd.apriori = 1, teta.unit = list("deg", "rad"))
}
\arguments{
\item{Qxy}{PARAM_DESCRIPTION}

\item{prob}{PARAM_DESCRIPTION, Default: NA}

\item{sd.apriori}{PARAM_DESCRIPTION, Default: 1}

\item{teta.unit}{PARAM_DESCRIPTION, Default: list("deg", "rad")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
