% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deformation_analysis_functions.r
\name{f_test}
\alias{f_test}
\title{f-test}
\usage{
f_test(sd1, sd2, df1, df2, prob)
}
\arguments{
\item{sd1}{First standard deviation}

\item{sd2}{Second standard deviation}

\item{df1}{First degrees of freedom}

\item{df2}{Second degrees of freedom}

\item{prob}{Probability}
}
\value{
Test decision
}
\description{
f-test
}
