% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{fdir_st}
\alias{fdir_st}
\title{FUNCTION_TITLE}
\usage{
fdir_st(st.survey.net, units.dir = "sec")
}
\arguments{
\item{st.survey.net}{PARAM_DESCRIPTION}

\item{units.dir}{PARAM_DESCRIPTION, Default: 'sec'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{arrange}}
}
