\name{firstSignif}
\alias{firstSignif}
\title{Finds first significant digit.}
\usage{
  firstSignif(x)
}
\arguments{
  \item{x}{Numeric vector}
}
\description{
  Finds first significant digit
}
\seealso{
  Other Formatting functions: \code{\link{formatPercent}},
  \code{\link{formatPound}}, \code{\link{formatRound}},
  \code{\link{roundFirstSignif}}
}

