% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{fix_params2D}
\alias{fix_params2D}
\title{FUNCTION_TITLE}
\usage{
fix_params2D(net.points)
}
\arguments{
\item{net.points}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
