% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{fmat}
\alias{fmat}
\title{FUNCTION_TITLE}
\usage{
fmat(survey.net, units.dir = "sec", units.dist = "mm")
}
\arguments{
\item{survey.net}{PARAM_DESCRIPTION}

\item{units.dir}{PARAM_DESCRIPTION, Default: 'sec'}

\item{units.dist}{PARAM_DESCRIPTION, Default: 'mm'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}},\code{\link[dplyr]{mutate}}
}
