% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{fmat1D}
\alias{fmat1D}
\title{FUNCTION_TITLE}
\usage{
fmat1D(survey.net, units = units)
}
\arguments{
\item{survey.net}{PARAM_DESCRIPTION}

\item{units}{PARAM_DESCRIPTION, Default: units}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
