\name{format_round}
\alias{format_round}
\title{Rounds to specified level of precision.}
\usage{format_round(x, n=1)}
\description{Rounds to specified level of precision.}
\details{By default, rounds to one digit.}
\keyword{internal}
\arguments{\item{x}{Numeric vector}
\item{n}{Number of digits, defaults to 1}}
