\name{format_values}
\alias{format_values}
\title{Applies format function to x.}
\usage{format_values(x, formatter)}
\description{Applies format function to x.}
\details{Applies format function (specified by formatter) to x.}
\keyword{internal}
\arguments{\item{x}{Character vector}
\item{formatter}{Formatting function}}
