\name{formatLabels}
\alias{formatLabels}
\title{Applies formatting to labels and calculates justification position.}
\usage{
  formatLabels(s)
}
\arguments{
  \item{s}{A surveyorStats object}
}
\description{
  Takes a surveyorStats object and adds two additional data
  columns: labelsValue and labelsJust.
}
\keyword{internal}

