\name{formatPercent}
\alias{formatPercent}
\title{Formats numeric value as percentage.}
\usage{
  formatPercent(x, digits = 1)
}
\arguments{
  \item{x}{Numeric vector}

  \item{digits}{Numeric. Number of digits.}
}
\description{
  Multiplies value by 100 and appends %% sign, e.g. 0.23 ->
  "23%"
}
\seealso{
  Other Formatting functions: \code{\link{firstSignif}},
  \code{\link{formatPound}}, \code{\link{formatRound}},
  \code{\link{roundFirstSignif}}
}

