\name{formatPound}
\alias{formatPound}
\title{Rounds to specified level of precision and prefixes a currency symbol.}
\usage{
  NA
}
\arguments{
  \item{x}{Numeric vector}

  \item{digits}{Passed to format}

  \item{nsmall}{Passed to format}

  \item{symbol}{Prepended to formatted string}
}
\description{
  Rounds to specified level of precision and prefixes a
  currency symbol.
}
\seealso{
  Other Formatting functions: \code{\link{firstSignif}},
  \code{\link{formatPercent}}, \code{\link{formatRound}},
  \code{\link{roundFirstSignif}}
}
\keyword{internal}

