\name{formatRound}
\alias{formatRound}
\title{Rounds to specified level of precision.}
\usage{
  formatRound(x, digits = 1)
}
\arguments{
  \item{x}{Numeric vector}

  \item{digits}{Number of digits, defaults to 1}
}
\description{
  By default, rounds to one digit.
}
\seealso{
  Other Formatting functions: \code{\link{firstSignif}},
  \code{\link{formatPercent}}, \code{\link{formatPound}},
  \code{\link{roundFirstSignif}}
}

