\name{formatValues}
\alias{formatValues}
\title{Applies format function to x.}
\usage{
  formatValues(x, formatter, ...)
}
\arguments{
  \item{x}{Character vector}

  \item{formatter}{Formatting function}

  \item{...}{Passed to formatting function}
}
\description{
  Applies format function (specified by formatter) to x.
}
\keyword{internal}

