\name{get_q_common_unique_pattern}
\alias{get_q_common_unique_pattern}
\title{Get common and unique text in question based on regex pattern identification...}
\usage{get_q_common_unique_pattern(x)}
\description{Get common and unique text in question based on regex pattern identification}
\arguments{\item{x}{A character vector}
\item{pattern}{A character vector with regex patterns}}
