\name{has_dk}
\alias{has_dk}
\title{Tests whether levels contain "Don't know"...}
\usage{has_dk(x, dk="Don't Know")}
\description{Tests whether levels contain "Don't know"}
\value{TRUE or FALSE}
\seealso{\code{\link{remove_dk}}, 
\code{\link{remove_all_dk}},}
\arguments{\item{x}{Vector or Factor}
\item{dk}{Character vector, containing search terms, e.g. c("Don't know", "Don't Know")}}
