\name{is.yesno}
\alias{is.yesno}
\title{Checks if question has response of yes/no.}
\usage{
  is.yesno(s)
}
\arguments{
  \item{s}{A surveyorStats object}
}
\description{
  Checks if question has response of yes/no.
}
\seealso{
  \code{\link{as.surveyorStats}}

  Other stats helper functions: \code{\link{allNA}},
  \code{\link{allNull}}, \code{\link{qType}},
  \code{\link{reorderQuestion}},
  \code{\link{reorderResponse}},
  \code{\link{splitBinCombine}},
  \code{\link{splitMeanCombine}},
  \code{\link{splitPercentCombine}},
  \code{\link{weightedCount}}, \code{\link{weightedMean}},
  \code{\link{weightedMedian}}, \code{\link{weightedSum}}
}
\keyword{internal}

