\name{latex_sideways}
\alias{latex_sideways}
\title{Prints a string to Latex, dealing with special characters...}
\usage{latex_sideways(x)}
\description{Prints a string to Latex, dealing with special characters}
\keyword{internal}
\keyword{Prepares}
\keyword{a}
\keyword{string}
\keyword{to}
\keyword{be}
\keyword{printed}
\keyword{sideways}
\keyword{(vertical)}
\keyword{in}
\keyword{LaTex}
\keyword{internal}
\arguments{\item{x}{A string containing question text}
\item{x}{A character vector}}
