% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{model_adequacy_test.shiny}
\alias{model_adequacy_test.shiny}
\title{FUNCTION_TITLE}
\usage{
model_adequacy_test.shiny(sd.apriori, sd.estimated, df, prob)
}
\arguments{
\item{sd.apriori}{PARAM_DESCRIPTION}

\item{sd.estimated}{PARAM_DESCRIPTION}

\item{df}{PARAM_DESCRIPTION}

\item{prob}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
