% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{ni}
\alias{ni}
\title{FUNCTION_TITLE}
\usage{
ni(pt1_coords, pt2_coords, type = list("dec", "dms", "rad"))
}
\arguments{
\item{pt1_coords}{PARAM_DESCRIPTION}

\item{pt2_coords}{PARAM_DESCRIPTION}

\item{type}{PARAM_DESCRIPTION, Default: list("dec", "dms", "rad")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
