\name{paste_percent}
\alias{paste_percent}
\title{Appends a percentage sign to numeric vector...}
\usage{paste_percent(x)}
\description{Appends a percentage sign to numeric vector}
\details{Turns a numeric vector into a character vector}
\keyword{internal}
\keyword{string}
\arguments{\item{x}{Numeric vector}}
