\name{plot_column}
\alias{plot_column}
\title{Plot data in column chart format}
\usage{
  plot_column(s, plot_function = "plot_column", ...)
}
\arguments{
  \item{s}{A surveyor_stats object}

  \item{plot_function}{Character vector: Identifies the
  name of the plot function used to create the plot}

  \item{...}{Ignored}
}
\description{
  Plot data in column chart format
}
\seealso{
  For an overview of the surveyor package
  \code{\link{surveyor}}

  Other plot_functions: \code{\link{plot_bar}},
  \code{\link{plot_bar_sum}}, \code{\link{plot_guess}},
  \code{\link{plot_net_score}}, \code{\link{plot_point}},
  \code{\link{plot_text}}
}
\keyword{plot}

