\name{plot_net_score}
\alias{plot_net_score}
\title{Plot data in net score format (bar chart, but percentage axis)...}
\usage{plot_net_score(f, surveyor)}
\description{Plot data in net score format (bar chart, but percentage axis)}
\seealso{plot_bar, plot_bar_stacked, plot_net_score, plot_point, plot_text}
\arguments{\item{f}{A data frame with coded answers, provided by a code_* function}
\item{surveyor}{Surveyor object}}
