\name{plot_point}
\alias{plot_point}
\title{Plot data in bubble chart format...}
\usage{plot_point(f, surveyor)}
\description{Plot data in bubble chart format}
\seealso{plot_bar, plot_bar_stacked, plot_net_score, plot_point, plot_text}
\arguments{\item{f}{A data frame with coded answers, provided by a code_* function}
\item{surveyor}{Surveyor object}}
