\name{plot_text}
\alias{plot_text}
\title{Plot data as text.}
\usage{
  plot_text(s, plot_function = "plot_text", ...)
}
\arguments{
  \item{s}{A surveyor_stats object}

  \item{plot_function}{Character vector: Identifies the
  name of the plot function used to create the plot}

  \item{...}{Ignored}
}
\description{
  Plots questions that are summarised using
  \code{\link{stats_text}}.
}
\seealso{
  Plot functions: \itemize{ \item \code{\link{plot_bar}}
  \item \code{\link{plot_bar_sum}} \item
  \code{\link{plot_column}} \item \code{\link{plot_point}}
  \item \code{\link{plot_text}} \item
  \code{\link{plot_net_score}} }

  For an overview of the surveyor package
  \code{\link{surveyor}}
}
\keyword{plot}

