\name{plot_yesno}
\alias{plot_yesno}
\title{Plot data in bar chart format, where the data consists of yes/no options.}
\usage{plot_yesno(s, surveyor)}
\description{Plot data in bar chart format, where the data consists of yes/no options.}
\details{This will produce a bar chart, but only "Yes" values will be plotted, i.e. the "No" values will be invisible.}
\seealso{Plot functions: 
\itemize{
\item \code{\link{plot_bar}} 
\item \code{\link{plot_point}} 
\item \code{\link{plot_text}} 
\item \code{\link{plot_net_score}} 
}

For an overview of the surveyor package \code{\link{surveyor}}}
\keyword{plot}
\arguments{\item{s}{A surveyor_stats object}
\item{surveyor}{Surveyor object}}
