\name{plotBar}
\alias{plotBar}
\title{Plot data in bar chart format}
\usage{
  plotBar(s, plotFunction = "plotBar", ...)
}
\arguments{
  \item{s}{A surveyorStats object}

  \item{plotFunction}{Character vector: Identifies the name
  of the plot function used to create the plot}

  \item{...}{ignored}
}
\description{
  Plot data in bar chart format
}
\seealso{
  For an overview of the surveyor package
  \code{\link{surveyor}}

  Other plotFunctions: \code{\link{plotBarSum}},
  \code{\link{plotColumn}}, \code{\link{plotGuess}},
  \code{\link{plotNetScore}}, \code{\link{plotPoint}},
  \code{\link{plotText}}
}
\keyword{plot}

