\name{plotColours}
\alias{plotColours}
\title{Apply Brewer pallete colours to plot}
\usage{
  plotColours(s, colours = 3,
    brewerPalette = s$surveyorDefaults$brewerPalette,
    revBrewerPal = s$surveyorDefaults$revBrewerPal, ...)
}
\description{
  Apply Brewer pallete colours to plot
}
\keyword{internal}

