\name{plotColumn}
\alias{plotColumn}
\title{Plot data in column chart format}
\usage{
  plotColumn(s, plotFunction = "plotColumn", ...)
}
\arguments{
  \item{s}{A surveyorStats object}

  \item{plotFunction}{Character vector: Identifies the name
  of the plot function used to create the plot}

  \item{...}{Ignored}
}
\description{
  Plot data in column chart format
}
\seealso{
  For an overview of the surveyor package
  \code{\link{surveyor}}

  Other plotFunctions: \code{\link{plotBar}},
  \code{\link{plotBarSum}}, \code{\link{plotGuess}},
  \code{\link{plotNetScore}}, \code{\link{plotNull}},
  \code{\link{plotPoint}}, \code{\link{plotText}}
}
\keyword{plot}

