\name{plotNull}
\alias{plotNull}
\title{Plots nothing, useful to suppress plot and return data only.}
\usage{
  plotNull(s, plotFunction = "plotNull", ...)
}
\arguments{
  \item{s}{A surveyorStats object}

  \item{plotFunction}{Character vector: Identifies the name
  of the plot function used to create the plot}

  \item{...}{ignored}
}
\description{
  Plots nothing, useful to suppress plot and return data
  only.
}
\seealso{
  For an overview of the surveyor package
  \code{\link{surveyor}}

  Other plotFunctions: \code{\link{plotBar}},
  \code{\link{plotBarSum}}, \code{\link{plotColumn}},
  \code{\link{plotGuess}}, \code{\link{plotNetScore}},
  \code{\link{plotPoint}}, \code{\link{plotText}}
}
\keyword{plot}

