\name{plotPoint}
\alias{plotPoint}
\title{Plot data in bubble chart format}
\usage{
  plotPoint(s, plotFunction = "plotPoint", ...)
}
\arguments{
  \item{s}{A surveyorStats object}

  \item{plotFunction}{Character vector: Identifies the name
  of the plot function used to create the plot}

  \item{...}{Ignored}
}
\description{
  Plot data in bubble chart format
}
\seealso{
  For an overview of the surveyor package
  \code{\link{surveyor}}

  Other plotFunctions: \code{\link{plotBar}},
  \code{\link{plotBarSum}}, \code{\link{plotColumn}},
  \code{\link{plotGuess}}, \code{\link{plotNetScore}},
  \code{\link{plotText}}
}
\keyword{plot}

