\name{print.ggplotmod}
\alias{print.ggplotmod}
\title{modified print method for ggplot to align title to plot instead of plotting grid.}
\usage{
  \method{print}{ggplotmod} (x, newpage = is.null(vp),
    vp = NULL, ...)
}
\arguments{
  \item{x}{plot to display}

  \item{newpage}{draw new (empty) page first?}

  \item{vp}{viewport to draw plot in}

  \item{...}{other arguments not used by this method}
}
\description{
  modified print method for ggplot to align title to plot
  instead of plotting grid.
}

