\name{print.surveyorPlot}
\alias{print.surveyorPlot}
\title{print method for surveyorPlot object.}
\usage{
  \method{print}{surveyorPlot} (x, ...)
}
\arguments{
  \item{x}{plot to display}

  \item{...}{other arguments not used by this method}
}
\description{
  print method for surveyorPlot object.
}

