\name{print_surveyor_question}
\alias{print_surveyor_question}
\title{Prints surveyor object output...}
\usage{print_surveyor_question(surveyor, q_id, counter, f, g, h, plot_size)}
\description{Prints surveyor object output}
\arguments{\item{surveyor}{A surveyor object}
\item{q_id}{The question id}
\item{counter}{The file number}
\item{f}{Results from code_* function}
\item{g}{Results from stats_* function}
\item{h}{Results from plot_* function}
\item{plot_size}{the plot size in inches}}
