\name{qType}
\alias{qType}
\title{Determines question type as single/grid question and single/multi response.}
\usage{
  qType(s)
}
\arguments{
  \item{s}{A surveyorStats object}
}
\description{
  Determines question type as single/grid question and
  single/multi response.
}
\seealso{
  \code{\link{as.surveyorStats}}

  Other stats helper functions: \code{\link{allNA}},
  \code{\link{allNull}}, \code{\link{is.yesno}},
  \code{\link{reorderQuestion}},
  \code{\link{reorderResponse}},
  \code{\link{splitBinCombine}},
  \code{\link{splitMeanCombine}},
  \code{\link{splitPercentCombine}},
  \code{\link{weightedCount}}, \code{\link{weightedMean}},
  \code{\link{weightedMedian}}, \code{\link{weightedSum}}
}
\keyword{internal}

