\docType{data}
\name{quiet}
\alias{quiet}
\title{Blank axis titles and no legend}
\format{List of 3
 $ legend.position: chr "none"
 $ axis.title.x   :function (...)  
  ..- attr(*, "class")= chr "theme"
  ..- attr(*, "type")= chr "any"
  ..- attr(*, "call")= language theme_blank()
 $ axis.title.y   :function (...)  
  ..- attr(*, "class")= chr "theme"
  ..- attr(*, "type")= chr "any"
  ..- attr(*, "call")= language theme_blank()
 - attr(*, "class")= chr "options"}
\usage{
  quiet
}
\description{
  Blank axis titles and no legend
}
\keyword{datasets}
\keyword{internal}

