\docType{data}
\name{quiet_axes}
\alias{quiet_axes}
\title{Blank axis titles}
\format{List of 2
 $ axis.title.x:function (...)  
  ..- attr(*, "class")= chr "theme"
  ..- attr(*, "type")= chr "any"
  ..- attr(*, "call")= language theme_blank()
 $ axis.title.y:function (...)  
  ..- attr(*, "class")= chr "theme"
  ..- attr(*, "type")= chr "any"
  ..- attr(*, "call")= language theme_blank()
 - attr(*, "class")= chr "options"}
\usage{
  quiet_axes
}
\description{
  Blank axis titles
}
\keyword{datasets}
\keyword{internal}

