% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{Qxy}
\alias{Qxy}
\title{FUNCTION_TITLE}
\usage{
Qxy(Qx, fix)
}
\arguments{
\item{Qx}{PARAM_DESCRIPTION}

\item{fix}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
