% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{read_surveynet}
\alias{read_surveynet}
\title{Read surveynet data}
\usage{
read_surveynet(file, dest_crs = NA, axes = c("Easting", "Northing"))
}
\arguments{
\item{file}{PARAM_DESCRIPTION}

\item{dest_crs}{Destination coordinate system (EPSG), Default: NA}

\item{axes}{order of coordinate axes, Default: c("Easting", "Northing")}
}
\value{
surveynet object
}
\description{
Read raw data stored in pre-defined excel file.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readxl]{read_excel}}
\code{\link[sf]{st_as_sf}},\code{\link[sf]{sfc}},\code{\link[sf]{sf}},\code{\link[sf]{st_crs}}
\code{\link[dplyr]{rename}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{select}},\code{\link[dplyr]{filter}}
\code{\link[purrr]{when}}
}
