\name{remove_all_dk}
\alias{remove_all_dk}
\title{Removes "Don't know" from factor levels in data frame...}
\usage{remove_all_dk(x, dk=c("Don't Know", "Don't know", "Dont know", "DK",
    "NA"))}
\description{Removes "Don't know" from factor levels in data frame}
\value{A data frame}
\arguments{\item{x}{Vector or Factor}
\item{dk}{Character vector, containing search terms, e.g. c("Don't know", "Don't Know")}}
