\name{remove_dk}
\alias{remove_dk}
\title{Removes "Don't know" from levels...}
\usage{remove_dk(x, dk="Don't Know")}
\description{Removes "Don't know" from levels}
\value{A factor with "Dont know" removed}
\seealso{\code{\link{has_dk}}, 
\code{\link{remove_all_dk}},}
\arguments{\item{x}{Vector or Factor}
\item{dk}{Character vector, containing search terms, e.g. c("Don't know", "Don't Know")}}
