\name{reorder_question}
\alias{reorder_question}
\title{Sorts data...}
\usage{reorder_question(df, reverse=FALSE)}
\description{Sorts data.frame questions in descending order by value.}
\details{Sorts data.frame question in descending order by value}
\value{A data frame}
\keyword{internal}
\arguments{\item{df}{A data frame containing at least two columns: response and value}}
