\name{reorder_response}
\alias{reorder_response}
\title{Sorts data...}
\usage{reorder_response(df)}
\description{Sorts data.frame responses in descending order by value.}
\details{Sorts data.frame responses in descending order by value}
\value{A data frame}
\keyword{internal}
\arguments{\item{df}{A data frame containing at least two columns: response and value}}
