\name{reorderResponse}
\alias{reorderResponse}
\title{Sorts data.frame responses in descending order by value.}
\usage{
  reorderResponse(df)
}
\arguments{
  \item{df}{A data frame containing at least two columns:
  response and value}
}
\value{
  A data frame
}
\description{
  Sorts data.frame responses in descending order by value
}
\seealso{
  \code{\link{as.surveyorStats}}

  Other stats helper functions: \code{\link{allNA}},
  \code{\link{allNull}}, \code{\link{is.yesno}},
  \code{\link{qType}}, \code{\link{reorderQuestion}},
  \code{\link{splitBinCombine}},
  \code{\link{splitMeanCombine}},
  \code{\link{splitPercentCombine}},
  \code{\link{weightedCount}}, \code{\link{weightedMean}},
  \code{\link{weightedMedian}}, \code{\link{weightedSum}}
}
\keyword{internal}

