\name{roundFirstSignif}
\alias{roundFirstSignif}
\title{Rounds to level of first significant digit.}
\usage{
  roundFirstSignif(x, f = round)
}
\arguments{
  \item{x}{Numeric vector}

  \item{f}{Rounding function, e.g. round, trunc, floor or
  ceiling}
}
\description{
  Rounds to level of first significant digit
}
\seealso{
  Other Formatting functions: \code{\link{firstSignif}},
  \code{\link{formatPercent}}, \code{\link{formatPound}},
  \code{\link{formatRound}}
}

