\name{save_print_function}
\alias{save_print_function}
\title{Creates function and static variable used to number charts in report...}
\usage{save_print_function(x)}
\description{Creates function and static variable used to number charts in report}
\details{Must be initialiased, e.g. ggsprint <- save_print_function}
\arguments{\item{x}{Surveyor object}}
\examples{ggsprint <<- save_print_function()}
