% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{sf.ellipse}
\alias{sf.ellipse}
\title{FUNCTION_TITLE}
\usage{
sf.ellipse(ellipse.param, scale = 10)
}
\arguments{
\item{ellipse.param}{PARAM_DESCRIPTION}

\item{scale}{PARAM_DESCRIPTION, Default: 10}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[nngeo]{st_ellipse}}
}
