% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{sigma.xy}
\alias{sigma.xy}
\title{FUNCTION_TITLE}
\usage{
\method{sigma}{xy}(Qxy.mat, sd.apriori)
}
\arguments{
\item{Qxy.mat}{PARAM_DESCRIPTION}

\item{sd.apriori}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
