% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deformation_analysis_functions.r
\name{deform.snet}
\alias{deform.snet}
\title{deform.snet}
\usage{
deform.snet(
  snet_1_path,
  snet_2_path,
  sd.apriori,
  units = "mm",
  dim_type = "1D",
  prob = 0.95,
  wdh_model = list("n_dh", "sd_dh", "d_dh", "E")
)
}
\arguments{
\item{snet_1_path}{Path to the data of the first epoch of the measurements}

\item{snet_2_path}{Path to the data of the second epoch of the measurements}

\item{sd.apriori}{Apriori standard deviation}

\item{units}{Units of the results, Default: 'mm'}

\item{dim_type}{Dimension type of the network, Default: '1D'}

\item{prob}{Probability for statistical testing, Default: 0.95}

\item{wdh_model}{Weightening model for leveling measurements, Default: list("n_dh", "sd_dh", "d_dh", "E")}
}
\value{
dataframe with two columns: "Name" indicating the name of the point and "stable" indicating the stability status of the point (TRUE or FALSE)
}
\description{
Two-epochs deformation analysis based on congruency testing
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{arrange}},\code{\link[dplyr]{select}}
\code{\link[MASS]{ginv}}
}
