% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snet1D_epoch_1}
\alias{snet1D_epoch_1}
\title{Simulated measurements in 1D geodetic control network - first epoch}
\format{
surveynet object.
}
\usage{
snet1D_epoch_1
}
\description{
Simulated measurements in 1D geodetic control network - first epoch
}
\keyword{datasets}
