% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snet2D}
\alias{snet2D}
\title{Measurements of the 2D geodetic control network in}
\format{
surveynet object.
}
\usage{
snet2D
}
\description{
Measurenents of the horizontal angles and distances in the 2D
network, in the Grdelica gorge, Serbia.
}
\keyword{datasets}
