\name{stats_bin}
\alias{stats_bin}
\title{Calculates summary statistics...}
\usage{stats_bin(f)}
\description{Calculates summary statistics}
\details{Takes the result of a code_function, e.g. code_single(), and calculates
summary values, for direct plotting by a plot_function, e.g. plot_bar()}
\value{A data frame with three columns: crossbreak, variable, value}
\arguments{\item{f}{A data frame with four columns: variable, value, crossbreak, weight}}
