\name{stats_mean}
\alias{stats_mean}
\title{Calculates numeric sum...}
\usage{stats_mean(x)}
\description{Calculates numeric sum}
\details{Add description}
\value{A data frame with three columns: cbreak, variable, value}
\seealso{Stats functions:
\itemize{
\item \code{\link{stats_bin}} 
\item \code{\link{stats_rank}} 
\item \code{\link{stats_net_score}}
}

For an overview of the surveyor package \code{\link{surveyor}}}
\keyword{stats}
\arguments{\item{x}{A data frame with four columns: cbreak, question, response, weight}}
