\name{stats_net_score}
\alias{stats_net_score}
\title{Code survey data as net score...}
\usage{stats_net_score(x)}
\description{Code survey data as net score}
\details{Code survey data in net score form}
\value{data frame}
\seealso{\code{\link{stats_bin}}}
\arguments{\item{x}{A data frame with four columns: crossbreak, question, response, weight}}
