\name{str.surveyorPlot}
\alias{str.surveyorPlot}
\title{Display structure of surveyorPlot object.}
\usage{
  str.surveyorPlot(x, max.level = 2, ...)
}
\arguments{
  \item{x}{surveyorPlot object}

  \item{max.level}{Passed to argument \code{max.level} in
  \code{\link{str}}. Defaults to 2, to limit the amount of
  information returned by \code{ggplot} and \code{lattice}}

  \item{...}{Other arguments passed to \code{\link{str}}}
}
\description{
  Structure is displayed to maximum of 2 levels
}
\seealso{
  \code{\link{str}}
}

