\name{str_common_unique}
\alias{str_common_unique}
\title{Finds the common and unique elements in a character vector...}
\usage{str_common_unique(x)}
\description{Finds the common and unique elements in a character vector}
\arguments{\item{x}{Character vector
#'}}
\examples{#' q_string <- c("Q_1", "Q_2", "Q_3") 
#' str_common_unique(q_string)$common
#' str_common_unique(q_string)$unique}
