\docType{package}
\name{surveyor-package}
\alias{"surveyor-package"}
\alias{surveyor}
\alias{surveyor-package}
\title{Tools for analysing and reporting on survey objects.}
\description{
  Tools for analysing and reporting on survey objects.
}
\details{
  This is a collection of functions to make it easier to
  analyse and report on market survey data.

  \itemize{ \item Initialise data and create surveyor
  object: \code{\link{as.surveyor}} \item Define default
  settings : \code{\link{surveyorDefaults}} \item Plot
  question: \code{\link{surveyorPlot}} }

  Coding functions: \itemize{ \item \code{\link{codeGuess}}
  \item \code{\link{codeSingle}} \item
  \code{\link{codeArray}} \item \code{\link{codeText}} }

  Summarising functions: \itemize{ \item
  \code{\link{statsGuess}} \item \code{\link{statsBin}}
  \item \code{\link{statsBinPercent}} \item
  \code{\link{statsRank}} \item \code{\link{statsNetScore}}
  \item \code{\link{statsSum}} }

  Plot functions: \itemize{ \item \code{\link{plotBar}}
  \item \code{\link{plotColumn}} \item
  \code{\link{plotPoint}} \item \code{\link{plotText}}
  \item \code{\link{plotNetScore}} }
}
\author{
  Andrie de Vries \email{andrie.de.vries@pentalibra.com}
}
\seealso{
  \code{\link{as.surveyor}}
}
\keyword{package}

