\name{surveyor_heading}
\alias{surveyor_heading}
\title{Creates latex heading.}
\usage{surveyor_heading(surveyor, x, headinglevel="chapter", pagebreak=FALSE)}
\description{Creates latex heading.}
\details{Creates latex heading}
\keyword{internal}
\arguments{\item{surveyor}{A surveyor object}
\item{x}{A character vector}
\item{level}{Character vector corresponding to latex heading, e.g. Chapter, Section, etc.}
\item{pagebreak}{Forces a page break if TRUE}}
