\name{surveyor_print_question}
\alias{surveyor_print_question}
\title{Prints surveyor question.}
\usage{
  surveyor_print_question(surveyor, q_id, f, g, h,
  plot_size)
}
\arguments{
  \item{surveyor}{A surveyor object}

  \item{q_id}{The question id}

  \item{f}{Results from code_* function}

  \item{g}{Results from stats_* function}

  \item{h}{Results from plot_* function}

  \item{plot_size}{the plot size in inches}
}
\description{
  Prints surveyor question.
}
\keyword{internal}

