\name{surveyor_save_plot}
\alias{surveyor_save_plot}
\title{Saves surveyor plot to pdf.}
\usage{surveyor_save_plot(surveyor, h, width, height, filename)}
\description{Saves surveyor plot to pdf.}
\details{Saves surveyor plot to pdf.  Uses either ggsave() or pdf(), depending on the class of plot}
\keyword{internal}
\arguments{\item{surveyor}{A surveyor object}
\item{h}{A surveyor plot object}
\item{width}{Width in inches}
\item{heigh}{Heigh in inches}
\item{filename}{Filename without path}}
