\name{surveyor_stats}
\alias{surveyor_stats}
\title{Creates surveyor_stats object, used as input to plot...}
\usage{surveyor_stats(data, ylabel="Fraction of respondents",
    formatter="percent")}
\description{Creates surveyor_stats object, used as input to plot}
\value{A surveyor_stats object}
\keyword{internal}
\arguments{\item{data}{A data frame}
\item{ylabel}{Character string to print as plot y label}}
