\name{surveyor_write}
\alias{surveyor_write}
\title{Writes result to surveyor sink file.}
\usage{surveyor_write(surveyor, x)}
\description{Writes result to surveyor sink file.}
\details{Appends text to surveyor sink file.}
\keyword{internal}
\arguments{\item{surveyor}{A surveyor object}
\item{x}{A character vector}}
