\name{surveyPlotApply}
\alias{surveyPlotApply}
\title{Apply surveyPlot to list of question definitions.}
\usage{
  surveyPlotApply(banner, s, mlist)
}
\arguments{
  \item{banner}{a list of arguments to send to
  \code{\link{surveyPlot}}}

  \item{s}{Optional surveyor object. If supplied, appends s
  to each element of defs}

  \item{mlist}{Options list of modified arguments}
}
\description{
  Apply surveyPlot to list of question definitions.
}
\seealso{
  Other Surveyor apply functions:
  \code{\link{surveyTableApply}}
}

