\docType{data}
\name{theme_minimal}
\alias{theme_minimal}
\title{Define minimal theme}
\format{List of 14
 $ axis.title.x     :function (...)  
  ..- attr(*, "class")= chr "theme"
  ..- attr(*, "type")= chr "any"
  ..- attr(*, "call")= language theme_blank()
 $ axis.title.y     :function (...)  
  ..- attr(*, "class")= chr "theme"
  ..- attr(*, "type")= chr "any"
  ..- attr(*, "call")= language theme_blank()
 $ axis.text.x      :function (...)  
  ..- attr(*, "class")= chr "theme"
  ..- attr(*, "type")= chr "any"
  ..- attr(*, "call")= language theme_blank()
 $ axis.text.y      :function (...)  
  ..- attr(*, "class")= chr "theme"
  ..- attr(*, "type")= chr "any"
  ..- attr(*, "call")= language theme_blank()
 $ axis.ticks       :function (...)  
  ..- attr(*, "class")= chr "theme"
  ..- attr(*, "type")= chr "any"
  ..- attr(*, "call")= language theme_blank()
 $ axis.ticks.margin:Class 'unit'  atomic [1:4] 0 0 0 0
  .. ..- attr(*, "unit")= chr "lines"
  .. ..- attr(*, "valid.unit")= int 3
 $ axis.ticks.length:Class 'unit'  atomic [1:1] 0
  .. ..- attr(*, "unit")= chr "cm"
  .. ..- attr(*, "valid.unit")= int 1
 $ panel.border     :function (...)  
  ..- attr(*, "class")= chr "theme"
  ..- attr(*, "type")= chr "any"
  ..- attr(*, "call")= language theme_blank()
 $ panel.ticks      :function (...)  
  ..- attr(*, "class")= chr "theme"
  ..- attr(*, "type")= chr "any"
  ..- attr(*, "call")= language theme_blank()
 $ panel.grid.major :function (...)  
  ..- attr(*, "class")= chr "theme"
  ..- attr(*, "type")= chr "any"
  ..- attr(*, "call")= language theme_blank()
 $ panel.grid.minor :function (...)  
  ..- attr(*, "class")= chr "theme"
  ..- attr(*, "type")= chr "any"
  ..- attr(*, "call")= language theme_blank()
 $ plot.margin      :Class 'unit'  atomic [1:4] 0 0 0 0
  .. ..- attr(*, "unit")= chr "lines"
  .. ..- attr(*, "valid.unit")= int 3
 $ panel.margin     :Class 'unit'  atomic [1:4] 0 0 0 0
  .. ..- attr(*, "unit")= chr "lines"
  .. ..- attr(*, "valid.unit")= int 3
 $ legend.position  : chr "none"
 - attr(*, "class")= chr "options"}
\usage{
  theme_minimal
}
\description{
  Define minimal theme
}
\keyword{datasets}
\keyword{internal}

