\name{theme_surveyor}
\alias{theme_surveyor}
\title{Sets up default surveyor theme for use in ggplot.}
\usage{theme_surveyor(base_size=12, base_family="")}
\description{Sets up default surveyor theme for use in ggplot.}
\keyword{internal}
\arguments{\item{surveyor}{A surveyor object}
\item{q_id}{The question id}
\item{counter}{The file number}
\item{f}{Results from code_* function}
\item{g}{Results from stats_* function}
\item{h}{Results from plot_* function}
\item{plot_size}{the plot size in inches}}
