% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{Wmat}
\alias{Wmat}
\title{FUNCTION_TITLE}
\usage{
Wmat(survey.net, sd.apriori = 1, res.units = "mm")
}
\arguments{
\item{survey.net}{PARAM_DESCRIPTION}

\item{sd.apriori}{PARAM_DESCRIPTION, Default: 1}

\item{res.units}{PARAM_DESCRIPTION, Default: 'mm'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}},\code{\link[dplyr]{select}},\code{\link[dplyr]{mutate}}
}
