% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{Wmat1D}
\alias{Wmat1D}
\title{FUNCTION_TITLE}
\usage{
Wmat1D(
  survey.net,
  wdh_model = list("sd_dh", "d_dh", "n_dh", "E"),
  sd0 = 1,
  d0 = NA,
  n0 = 1,
  res.units = "mm"
)
}
\arguments{
\item{survey.net}{PARAM_DESCRIPTION}

\item{wdh_model}{PARAM_DESCRIPTION, Default: list("sd_dh", "d_dh", "n_dh", "E")}

\item{sd0}{PARAM_DESCRIPTION, Default: 1}

\item{d0}{PARAM_DESCRIPTION, Default: NA}

\item{n0}{PARAM_DESCRIPTION, Default: 1}

\item{res.units}{PARAM_DESCRIPTION, Default: 'mm'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}
}
