\name{code_array}
\alias{code_array}
\title{Code survey data in array form...}
\usage{code_array(surveyor, Qid, multicode=FALSE, remove.other=FALSE,
    index="crossbreak", weight=TRUE, ...)}
\description{Code survey data in array form}
\seealso{code_single, code_rank}
\arguments{\item{surveyor}{Surveyor object}
\item{Qid}{Question id}
\item{multicode}{Boolean}
\item{remove.other}{Boolean}
\item{index}{Crossbreak variable}
\item{weight}{}
\item{...}{}}
