\name{code_array}
\alias{code_array}
\title{Code survey data in array question form (i...}
\usage{code_array(surveyor, Qid, multicode=FALSE, remove.other=FALSE)}
\description{Code survey data in array question form (i.e. with subquestions)}
\seealso{code_single, code_rank}
\arguments{\item{surveyor}{Surveyor object}
\item{Qid}{Question id}
\item{multicode}{If true, will code for questions with multiple values}
\item{remove.other}{If true, will remove last column}}
