\name{code_guess}
\alias{code_guess}
\title{Code survey data in single or array question form}
\usage{
  code_guess(surveyor, q_id, ...)
}
\arguments{
  \item{surveyor}{Surveyor object}

  \item{q_id}{Question id}

  \item{...}{Other parameters passed on to downstream
  code_* functions}
}
\value{
  data frame
}
\description{
  Code survey data in array question form (i.e. with
  subquestions)
}
\seealso{
  For an overview of the surveyor package
  \code{\link{surveyor}}

  Other code_functions: \code{\link{code_array}},
  \code{\link{code_single}}, \code{\link{code_text}}
}
\keyword{code}

