\name{code_rank}
\alias{code_rank}
\title{Code survey data in rank question form...}
\usage{code_rank(surveyor, q_id, remove.other=FALSE)}
\description{Code survey data in rank question form}
\details{Code survey data in rank question form (i.e. with subquestions)}
\value{data frame}
\seealso{\code{\link{code_single}}, \code{\link{code_array}}}
\arguments{\item{surveyor}{Surveyor object}
\item{q_id}{Question id}
\item{remove.other}{Boolean}}
