\name{code_single}
\alias{code_single}
\title{Code survey data in array question form (i...}
\usage{code_single(surveyor, Qid)}
\description{Code survey data in array question form (i.e. with subquestions)}
\seealso{code_array, code_rank}
\arguments{\item{surveyor}{Surveyor object}
\item{Qid}{Question id}}
