\name{code_single}
\alias{code_single}
\title{Code survey data in single question form...}
\usage{code_single(surveyor, q_id, ...)}
\description{Code survey data in single question form}
\details{Code survey data in single question form (i.e. without subquestions)}
\seealso{Other coding functions: 
\code{\link{code_single}},  
\code{\link{code_array}}, 
\code{\link{code_text}}, 
\code{\link{code_guess}}

For an overview of the surveyor package \code{\link{surveyor}}}
\value{data frame}
\keyword{code}
\arguments{\item{surveyor}{Surveyor object}
\item{q_id}{Question id}
\item{...}{Not used}}
