\name{code_text}
\alias{code_text}
\title{Code survey data in text question form (i...}
\usage{code_text(surveyor, q_id, inn, out, filter_responses=TRUE)}
\description{Code survey data in text question form (i.e. open-ended text}
\seealso{Other coding functions: 
\code{\link{code_single}},  
\code{\link{code_array}}, 
\code{\link{code_text}}, 
\code{\link{code_guess}}

For an overview of the surveyor package \code{\link{surveyor}}}
\value{data frame}
\keyword{code}
\arguments{\item{surveyor}{Surveyor object}
\item{q_id}{Question id}
\item{inn}{string passed to latexTranslate()}
\item{out}{string passed to latexTranslate()}
\item{filter_responses}{Indicates whether responses should be filtered first}}
