\name{code_text}
\alias{code_text}
\title{Code survey data in text question form (i...}
\usage{code_text(surveyor, q_id, inn, out)}
\description{Code survey data in text question form (i.e. open-ended text}
\seealso{Coding functions:
\code{\link{code_single}}, 
\code{\link{code_array}},
\code{\link{code_text}}
Summarising functions:
\code{\link{stats_bin}}, 
\code{\link{stats_net_score}}
Plot functions: 
\code{\link{plot_bar}}, 
\code{\link{plot_point}}}
\value{data frame}
\arguments{\item{surveyor}{Surveyor object}
\item{q_id}{Question id}
\item{inn}{string passed to latexTranslate()}
\item{out}{string passed to latexTranslate()}}
