\name{format_labels}
\alias{format_labels}
\title{Applies formatting to labels and calculates justification position.}
\usage{
  format_labels(s)
}
\arguments{
  \item{s}{A surveyor_stats object}
}
\description{
  Takes a surveyor_stats object and adds two additional
  data columns: value_labels and labels_just.
}
\keyword{internal}

