\name{paste_percent}
\alias{paste_percent}
\title{Formats numeric value as percentage.}
\usage{
  paste_percent(x, digits = 1)
}
\arguments{
  \item{x}{Numeric vector}

  \item{digits}{Numeric. Number of digits.}
}
\description{
  Multiplies value by 100 and appends % sign, e.g. 0.23 ->
  "23%"
}
\keyword{string}

