\name{plot_bar}
\alias{plot_bar}
\title{Plot data in horizontal bar chart format...}
\usage{plot_bar(f, surveyor)}
\description{Plot data in horizontal bar chart format}
\seealso{plot_bar, plot_bar_stacked, plot_net_score, plot_point, plot_text}
\arguments{\item{f}{A data frame with coded answers, provided by a code_* function}
\item{surveyor}{Surveyor object}}
