\name{plot_bar}
\alias{plot_bar}
\title{Plot data in bar chart format...}
\usage{plot_bar(s, surveyor)}
\description{Plot data in bar chart format}
\seealso{Coding functions:
\code{\link{code_single}}, 
\code{\link{code_array}},
\code{\link{code_text}}
Summarising functions:
\code{\link{stats_bin}}, 
\code{\link{stats_net_score}}
Plot functions: 
\code{\link{plot_bar}}, 
\code{\link{plot_point}}}
\arguments{\item{s}{A surveyor_stats object}
\item{surveyor}{Surveyor object}}
