\name{plot_bar}
\alias{plot_bar}
\title{Plot data in bar chart format}
\usage{
  plot_bar(s, plot_function = "plot_bar", ...)
}
\arguments{
  \item{s}{A surveyor_stats object}

  \item{plot_function}{Character vector: Identifies the
  name of the plot function used to create the plot}

  \item{...}{ignored}
}
\description{
  Plot data in bar chart format
}
\seealso{
  Plot functions: \itemize{ \item \code{\link{plot_bar}}
  \item \code{\link{plot_bar_sum}} \item
  \code{\link{plot_column}} \item \code{\link{plot_point}}
  \item \code{\link{plot_text}} \item
  \code{\link{plot_net_score}} }

  For an overview of the surveyor package
  \code{\link{surveyor}}
}
\keyword{plot}

