\name{plot_bar_sum}
\alias{plot_bar_sum}
\title{Plot data in bar chart format without modifying format.}
\usage{plot_bar_sum(s, surveyor)}
\description{Plot data in bar chart format without modifying format.}
\details{The standard plot_bar() function will plot the data in a stacked bar chart format and apply percentage formatting.  plot_bar_sum() applies no formatting.}
\seealso{Plot functions: 
\itemize{
\item \code{\link{plot_bar}} 
\item \code{\link{plot_bar_sum}} 
\item \code{\link{plot_column}} 
\item \code{\link{plot_point}} 
\item \code{\link{plot_text}} 
\item \code{\link{plot_net_score}} 
}

For an overview of the surveyor package \code{\link{surveyor-package}}}
\keyword{plot}
\arguments{\item{s}{A surveyor_stats object}
\item{surveyor}{Surveyor object}}
