\name{plot_column}
\alias{plot_column}
\title{Plot data in column chart format...}
\usage{plot_column(s, surveyor)}
\description{Plot data in column chart format}
\seealso{Plot functions: 
\itemize{
\item \code{\link{plot_bar}} 
\item \code{\link{plot_point}} 
\item \code{\link{plot_text}} 
\item \code{\link{plot_net_score}} 
}

For an overview of the surveyor package \code{\link{surveyor}}}
\keyword{plot}
\arguments{\item{s}{A surveyor_stats object}
\item{surveyor}{Surveyor object}}
