\name{plot_net_score}
\alias{plot_net_score}
\title{Plot data in net score format (bar chart, but percentage axis)...}
\usage{plot_net_score(s, surveyor)}
\description{Plot data in net score format (bar chart, but percentage axis)}
\seealso{Plot functions: 
\itemize{
\item \code{\link{plot_bar}} 
\item \code{\link{plot_point}} 
\item \code{\link{plot_text}} 
\item \code{\link{plot_net_score}} 
}

For an overview of the surveyor package \code{\link{surveyor}}}
\keyword{plot}
\arguments{\item{s}{A data frame with coded answers, provided by a code_* function}
\item{surveyor}{Surveyor object}}
