\name{plot_net_score}
\alias{plot_net_score}
\title{Plot data in net score format (bar chart, but percentage axis)}
\usage{
  plot_net_score(s, plot_function = "plot_net_score", ...)
}
\arguments{
  \item{s}{A surveyor_stats object}

  \item{plot_function}{Character vector: Identifies the
  name of the plot function used to create the plot}

  \item{...}{Ignored}
}
\description{
  Plot data in net score format (bar chart, but percentage
  axis)
}
\seealso{
  Plot functions: \itemize{ \item \code{\link{plot_bar}}
  \item \code{\link{plot_bar_sum}} \item
  \code{\link{plot_column}} \item \code{\link{plot_point}}
  \item \code{\link{plot_text}} \item
  \code{\link{plot_net_score}} }

  For an overview of the surveyor package
  \code{\link{surveyor}}
}
\keyword{plot}

