\name{plot_text}
\alias{plot_text}
\title{Plot data as text...}
\usage{plot_text(f, surveyor)}
\description{Plot data as text}
\seealso{Plot functions: 
\itemize{
\item \code{\link{plot_bar}} 
\item \code{\link{plot_bar_sum}} 
\item \code{\link{plot_column}} 
\item \code{\link{plot_point}} 
\item \code{\link{plot_text}} 
\item \code{\link{plot_net_score}} 
}

For an overview of the surveyor package \code{\link{surveyor-package}}}
\keyword{plot}
\arguments{\item{f}{A data frame with coded answers, provided by a code_* function}
\item{surveyor}{Surveyor object}}
