\name{plot_text}
\alias{plot_text}
\title{Plot data as text.}
\usage{plot_text(s, surveyor)}
\description{Plot data as text.}
\details{Plots questions that are summarised using \code{\link{stats_text}}.}
\seealso{Plot functions: 
\itemize{
\item \code{\link{plot_bar}} 
\item \code{\link{plot_bar_sum}} 
\item \code{\link{plot_column}} 
\item \code{\link{plot_point}} 
\item \code{\link{plot_text}} 
\item \code{\link{plot_net_score}} 
}

For an overview of the surveyor package \code{\link{surveyor}}}
\keyword{plot}
\arguments{\item{s}{A surveyor_stats object}
\item{surveyor}{Surveyor object}}
