\name{plot_text}
\alias{plot_text}
\title{Plot data as text.}
\usage{
  plot_text(s, plot_function = "plot_text", ...)
}
\arguments{
  \item{s}{A surveyor_stats object}

  \item{plot_function}{Character vector: Identifies the
  name of the plot function used to create the plot}

  \item{...}{Ignored}
}
\description{
  Plots questions that are summarised using
  \code{\link{stats_text}}.
}
\seealso{
  For an overview of the surveyor package
  \code{\link{surveyor}}

  Other plot_functions: \code{\link{plot_bar}},
  \code{\link{plot_bar_sum}}, \code{\link{plot_column}},
  \code{\link{plot_guess}}, \code{\link{plot_net_score}},
  \code{\link{plot_point}}
}
\keyword{plot}

