\name{reorder_response}
\alias{reorder_response}
\title{Sorts data.frame responses in descending order by value.}
\usage{
  reorder_response(df)
}
\arguments{
  \item{df}{A data frame containing at least two columns:
  response and value}
}
\value{
  A data frame
}
\description{
  Sorts data.frame responses in descending order by value
}
\keyword{internal}

