\name{reorderQuestion}
\alias{reorderQuestion}
\title{Sorts data.frame questions in descending order by value.}
\usage{
  reorderQuestion(df, reverse = FALSE)
}
\arguments{
  \item{df}{A data frame containing at least two columns:
  response and value}
}
\value{
  A data frame
}
\description{
  Sorts data.frame question in descending order by value
}
\keyword{internal}

