\name{round_first_signif}
\alias{round_first_signif}
\title{Rounds to level of first significant digit.}
\usage{
  round_first_signif(x, f = round)
}
\arguments{
  \item{x}{Numeric vector}

  \item{f}{Rounding function, e.g. round, trunc, floor or
  ceiling}
}
\description{
  Rounds to level of first significant digit
}
\keyword{internal}

