\name{stats_mean}
\alias{stats_mean}
\title{Calculates numeric sum...}
\usage{stats_mean(surveyor_code)}
\description{Calculates numeric sum}
\details{Add description}
\value{A data frame with three columns: cbreak, variable, value}
\seealso{Stats functions:
\itemize{
\item \code{\link{stats_bin}} 
\item \code{\link{stats_rank}} 
\item \code{\link{stats_net_score}}
}

For an overview of the surveyor package \code{\link{surveyor}}}
\keyword{stats}
\arguments{\item{surveyor_code}{An object of class "surveyor_code".  This is a list with the first element being a data frame with four columns: cbreak, question, response, weight}}
