\name{stats_net_score}
\alias{stats_net_score}
\title{Code survey data as net score...}
\usage{stats_net_score(f)}
\description{Code survey data as net score}
\details{Code survey data in net score form}
\value{data frame}
\seealso{\code{\link{code_single}}, \code{\link{code_rank}}}
\arguments{\item{f}{Output from a code_function()}}
