\name{stats_net_score}
\alias{stats_net_score}
\title{Code survey data as net score.}
\usage{stats_net_score(x)}
\description{Code survey data as net score.}
\details{Code survey data in net score form}
\value{data frame}
\seealso{Stats functions:
\itemize{
\item \code{\link{stats_bin}} 
\item \code{\link{stats_rank}} 
\item \code{\link{stats_net_score}}
}

For an overview of the surveyor package \code{\link{surveyor}}}
\keyword{stats}
\arguments{\item{x}{A data frame with four columns: cbreak, question, response, weight}}
