\name{stats_text}
\alias{stats_text}
\title{Code survey data as text.}
\usage{stats_text(x)}
\description{Code survey data as text.}
\details{Code survey data in text form}
\value{data frame}
\seealso{Stats functions:
\itemize{
\item \code{\link{stats_bin}} 
\item \code{\link{stats_rank}} 
\item \code{\link{stats_net_score}}
}

For an overview of the surveyor package \code{\link{surveyor}}}
\keyword{stats}
\arguments{\item{x}{A data frame column}}
