\name{surveyorPrintQuestion}
\alias{surveyorPrintQuestion}
\title{Prints surveyor question.}
\usage{
  surveyorPrintQuestion(surveyor, q_id, f, g, h, plotSize)
}
\arguments{
  \item{surveyor}{A surveyor object}

  \item{q_id}{The question id}

  \item{f}{Results from code* function}

  \item{g}{Results from stats* function}

  \item{h}{Results from plot* function}

  \item{plotSize}{the plot size in inches}
}
\description{
  Prints surveyor question.
}
\keyword{internal}

