\name{table_guess}
\alias{table_guess}
\title{Prints cbreak information in latex format...}
\usage{table_guess(g)}
\description{Prints cbreak information in latex format}
\details{Prints cbreak information in latex format}
\seealso{Coding functions:
\code{\link{code_single}}, 
\code{\link{code_array}},
\code{\link{code_text}}
Summarising functions:
\code{\link{stats_bin}}, 
\code{\link{stats_net_score}}
Plot functions: 
\code{\link{plot_bar}}, 
\code{\link{plot_point}}}
\value{data frame}
\keyword{internal}
\arguments{\item{g}{A surveyor_stats object}}
